refactor: getMyProfileId.ts => getUserInfo.ts 로 수정
